requirements.txt
Flask==3.0.3 
google-generativeai==0.8.3 
markdown==3.7 
python-dotenv==1.0.1